requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
PyJWT==2.8.0
requests==2.31.0
